\name{hello}
\alias{hello}
\title{Hello, World!}
\usage{
hello()
}
\description{
Prints 'Hello, world!'.
}
\examples{
hello()
}
\name{mergepairing}
\alias{mergepairing}
\title{Merge pairing test}
\usage{
mergepairing()
}
\description{
Runs merge pairing algorithm on the mergepairtesting file
}
\examples{
mergepairing()
}
