% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReebGraphPairing-package.R
\docType{package}
\name{ReebGraphPairing-package}
\alias{ReebGraphPairing}
\alias{ReebGraphPairing-package}
\title{ReebGraphPairing: Title: Interface to Java-Based ReebGraphPairing World Functionality Using rJava}
\description{
This is a simple example package that prints 'Hello, world!' to the console.
}
\author{
\strong{Maintainer}: Satyajit Mohanty \email{mohantysatyajit57@gmail.com}

Authors:
\itemize{
  \item Shubham Singh \email{shubhamus16091998@example.com}
}

}
\keyword{internal}
