% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as-reeb-graph.R
\name{as_reeb_graph}
\alias{as_reeb_graph}
\title{Coerce objects to class \code{reeb_graph}}
\usage{
as_reeb_graph(x, ...)
}
\arguments{
\item{x}{An R object to be coerced. See Details.}

\item{...}{Additional arguments passed to methods.}

\item{values}{A character value or a numeric vector; the node attribute to
use as the Reeb graph value function, or the values in order of node
indices. If \code{NULL} (the default), the first numeric node attribute is used.}
}
\value{
A \link{reeb_graph} object.
}
\description{
Coerce objects to \link{reeb_graph}-class objects.
}
\details{
The \code{as_reeb_graph()} methods require a network (mathematical graph)
structure and a real-valued function on the vertex set.
}
\examples{
library(igraph)
( g <- make_kautz_graph(2, 1) )
( l <- layout_nicely(g) )
plot(g, layout = l)
( rg <- as_reeb_graph(g, l[, 1]) )
# TODO: Coerce back to 'igraph' class and plot using Sugiyama layout.

}
